// src/app/Model/Product.ts
export class Product {
      id: number ;
      nameP: string;
      descriptionP: string ;
      imageP: string;
      fileVector : string;
      filePDF : string;
      filePNG : string;
      fileIcons : string;
      fileSVG : string;
      fileEPS : string;
      formatP: string;
      styleGraphicP: string;
      colorP: string;

    }